ResizeWindow refactor 1
